requirements.txt
python-docx==1.1.2
PyPDF2==3.0.1
matplotlib==3.9.2
Flask==3.0.3
